HistoricoFilmes/requirements.txt
Flask==3.1.1
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
Flask-Cors==4.0.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41